ik_fit_func is now a member function; random_joints moved to utilities module
